Don't wait for instances to be healthy when removing ASGs
If `wait_for_instances` is unset or set to True, then the instances
need to leave Pending:Wait before they're terminated. If there's a
problem with the logic that triggers the lifecycle hook, this might
not ever happen. If we're removing ASGs, we don't care about the
health of the instances